fix race condition in case of failed ip update
Currently when the curl request to the inwx API fails, the script
finishes and never tries again that day to update the DNS record
although it failed.

This commit moves the line where the recent ip gets written into a text
file after the case where everything went wrong. Thus in case the
script finishes with `exit 1` the ip file does not get written. That
means in the next run it is tried again to update the DNS
record with the most recent ip.