Fix race condition with multiple domain records
This commit fixes a race condition, where only the first inwx id of an
array of multiple ids gets updated. The section to save the old v4 and
v6 address is moved out of the for loop in order to have these variables
saved globally. This way we have still the old ip address saved, even in
the second run of the for loop.

This commit is a similar solution to the more complex solutions in #23
and #25